small fixes to make compound loss work

+++ transformer_seg/dataset.py
@@ -222,7 +222,7 @@ class BaselineSegmentationDataset(Dataset):
         line_cls = torch.full((len(lines),), self.line_token_id-1, dtype=torch.long)
         line_cls[0] = self.bos_token_id-1
         line_cls[-1] = self.eos_token_id-1
-        line_cls = F.one_hot(line_cls, num_classes=3)
+        line_cls = F.one_hot(line_cls, num_classes=3).float()
         return {'image': im,
                 'tokens': line_cls,
                 'curves': lines}
