+++ transformer_seg/modules/transformer.py
@@ -646,7 +646,7 @@ class TransformerDecoder(nn.Module):
             output = self.chunked_output(h)
         else:
             # shape: [b, seq_len, out_dim]
-            output = self.output(h).float()
+            output = self.output(h)
 
         # Output list if hidden states are requested, otherwise just the output
         # TODO: always output a list to have a consistent output type
